app/requirements.txt
fastapi
uvicorn
requests
pandas
openpyxl
sqlalchemy
databases
jinja2
python-multipart
bcrypt
passlib[bcrypt]
itsdangerous
python-jose
